deps: the python-minor-and-patch group, and the type errors numpy 2.5 exposed
Dependabot #22, 20 packages. Named "minor and patch" but it is not a
small change: fastapi 0.115->0.141, uvicorn 0.34->0.52, numpy 2.2->2.5,
scipy 1.15->1.18, ai-edge-litert 2.1.6->2.2.0, ruff 0.8.6->0.16.2.

Two things actually broke, and both were worth finding.

**click 8.4 removed CliRunner(mix_stderr=...)**, so six CLI test modules
failed to collect. Separate stdout/stderr is now the default, which is
exactly what mix_stderr=False was asking for, so the fix is to drop the
argument. 46 CLI tests pass.

**numpy 2.5's stricter stubs broke the type gate**, at mypy 1.14.1, which
had been clean. 16 errors, and they were not all noise:

  - `DetectorHealth.state` is a Literal of ok/degraded/unavailable/error
    while `DetectorWorker` sets "starting" and "stopped" -- so the API
    could publish a state its own contract said was impossible, and did,
    on every startup. Nothing enforces a Literal at runtime, which is why
    it went unnoticed. Widened to match reality (new `DetectorState`
    alias) rather than narrowed to hide it: a detector that is starting
    or stopped is a real thing an operator should see.
  - birdnet's `_probabilities` and activity's `_noise_floor_db` were
    declared `| None` and returned where the caller's annotation promised
    a bare ndarray.
  - `_timezone` inferred its type from the UTC fallback, so the ZoneInfo
    assignment on the happy path was the error.
  - `bucket_expression` took `ColumnElement` while every caller passes an
    `InstrumentedAttribute`, which is not a subclass of it. `history.py`
    already had `SourceKindColumn` for precisely this, with a comment
    saying it had been wrong at every call site once before. Reused it.

ruff 0.16.2 adds RUF046/RUF059 (32 findings, all mechanical) and one
ASYNC240: a blocking `stat()` in an async CLI command, sitting between a
`mkdir` and an `sf.write` the rule does not flag. Kept with a noqa and
the reason -- that is `oo audio`, not `oo serve`, so there is no capture
loop in the process for it to starve.

RUF046 removes `int()` around `round()`/`math.ceil()` in the audio hot
path, which is only safe if no operand is a numpy scalar --
`round(np.float64)` could have returned a numpy scalar used as an array
index. Checked rather than assumed: numpy's `__round__` returns a Python
int when called without ndigits.

Verified: ruff clean, mypy clean, 967 passed / 12 skipped, and the
resampler still bit-identical (impulse peak 14525, energy
0.014852375242005115, alias 3.7e-09, deficit 762).

NOT yet verified: ai-edge-litert 2.2.0. The BirdNET fixture test skips on
this laptop (no model assets), so nothing here has exercised it, and
CLAUDE.md is explicit that a detector counts as supported only after a
fixture test passes on the target architecture. That runs on the Pi next.

+++ src/open_observatory/audio/alsa_source.py
@@ -175,7 +175,7 @@ class AlsaSource:
         real loss between 0.10 s and 0.24 s, i.e. about one block each.
         """
         wanted = max(self._buffer_ms, 2.0 * self._block_ms)
-        return max(8, int(math.ceil(wanted / self._period_ms)))
+        return max(8, math.ceil(wanted / self._period_ms))
 
     async def open(self) -> StreamInfo:
         return await asyncio.get_running_loop().run_in_executor(
@@ -211,7 +211,7 @@ class AlsaSource:
             if fmt_const is None or sample_format not in _NUMPY_DTYPE:
                 continue
             for rate in candidates:
-                period_frames = max(64, int(round(rate * self._period_ms / 1000.0)))
+                period_frames = max(64, round(rate * self._period_ms / 1000.0))
                 try:
                     pcm = alsaaudio.PCM(
                         alsaaudio.PCM_CAPTURE,
@@ -252,7 +252,7 @@ class AlsaSource:
         self._scale = _SCALE[sample_format]
         self._period_frames = int(negotiated.get("period_size", period_frames) or period_frames)
         self._block_frames = max(
-            self._period_frames, int(round(rate * self._block_ms / 1000.0))
+            self._period_frames, round(rate * self._block_ms / 1000.0)
         )
         # Round the block to a whole number of periods so reads align.
         periods_per_block = max(1, self._block_frames // self._period_frames)
